need to make large object structure consistent with other postgres variable length structures.  Woudln't work on a big-endian 64 bit machine otherwise.

+++ src/backend/utils/large_object.h
@@ -61,7 +61,7 @@ typedef struct
  * Attribute by the Postgres access methods.
  */
 
-	long lo_length;
+	int32 lo_length;
 
 /*
  * Large object storage type - described above.
